NMS-13547: refactor passive checks
