--- core/ipc/rpc/common/src/main/java/org/opennms/core/rpc/echo/EchoRpcHealthCheck.java
@@ -1,75 +1,0 @@
-/*******************************************************************************
- * This file is part of OpenNMS(R).
- *
- * Copyright (C) 2018-2021 The OpenNMS Group, Inc.
- * OpenNMS(R) is Copyright (C) 1999-2021 The OpenNMS Group, Inc.
- *
- * OpenNMS(R) is a registered trademark of The OpenNMS Group, Inc.
- *
- * OpenNMS(R) is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published
- * by the Free Software Foundation, either version 3 of the License,
- * or (at your option) any later version.
- *
- * OpenNMS(R) is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with OpenNMS(R).  If not, see:
- *      http://www.gnu.org/licenses/
- *
- * For more information contact:
- *     OpenNMS(R) Licensing <license@opennms.org>
- *     http://www.opennms.org/
- *     http://www.opennms.com/
- *******************************************************************************/
-
-package org.opennms.core.rpc.echo;
-
-import static org.opennms.core.health.api.HealthCheckConstants.ECHO;
-
-import java.util.Arrays;
-import java.util.List;
-
-import org.opennms.core.health.api.Context;
-import org.opennms.core.health.api.HealthCheck;
-import org.opennms.core.health.api.Response;
-import org.opennms.core.health.api.Status;
-import org.opennms.core.health.impl.DefaultCachingHealthCheck;
-
-/**
- * Checks if the echo rpc module was called.
- * <p>
- * This is a passive health check, i.e. it never does some communication by itself. It can be wired with
- * the {@link EchoRpcModule} in order to get informed of echo rpcs.
- */
-public class EchoRpcHealthCheck extends DefaultCachingHealthCheck {
-
-    public EchoRpcHealthCheck() {
-        super(
-                new HealthCheck() {
-                    @Override
-                    public String getDescription() {
-                        return "Echo RPC";
-                    }
-
-                    @Override
-                    public List<String> getTags() {
-                        return Arrays.asList(ECHO);
-                    }
-
-                    @Override
-                    public Response perform(Context context) throws Exception {
-                        if (context.getMaxAge().isZero()) {
-                            return new Response(Status.Failure, "the 'maxAge' property of the health check context must be non-zero for the echo rpc health check");
-                        } else {
-                            return new Response(Status.Failure, String.format("Did not receive a recent echo rpc call - maxAge: %s", context.getMaxAge()));
-                        }
-                    }
-                }
-        );
-    }
-
-}
