+++ features/minion/core/health-check/src/main/resources/OSGI-INF/blueprint/blueprint.xml
@@ -1,0 +1,51 @@
+<blueprint
+    xmlns="http://www.osgi.org/xmlns/blueprint/v1.0.0"
+    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+    xsi:schemaLocation="
+        http://www.osgi.org/xmlns/blueprint/v1.0.0 
+        https://osgi.org/xmlns/blueprint/v1.0.0/blueprint.xsd
+">
+
+    <bean id="passiveEchoRpcHealthCheck" class="org.opennms.core.health.api.DefaultPassiveHealthCheck">
+        <argument value="Echo RPC (passive)"/>
+        <argument>
+            <list>
+                <value>broker</value>
+                <value>passive</value>
+            </list>
+        </argument>
+        <argument value="Unknown"/>
+    </bean>
+
+    <bean id="passiveRestConnectionHealthCheck" class="org.opennms.core.health.api.DefaultPassiveHealthCheck">
+        <argument value="Connecting to OpenNMS ReST API (passive)"/>
+        <argument>
+            <list>
+                <value>rest</value>
+                <value>passive</value>
+            </list>
+        </argument>
+        <argument value="Unknown"/>
+    </bean>
+
+    <service ref="passiveEchoRpcHealthCheck">
+        <interfaces>
+            <value>org.opennms.core.health.api.HealthCheck</value>
+            <value>org.opennms.core.health.api.HealthCheckResponseCache</value>
+        </interfaces>
+        <service-properties>
+            <entry key="alias" value="opennms.passiveEchoRpcHealthCheck"/>
+        </service-properties>
+    </service>
+
+    <service ref="passiveRestConnectionHealthCheck">
+        <interfaces>
+            <value>org.opennms.core.health.api.HealthCheck</value>
+            <value>org.opennms.core.health.api.HealthCheckResponseCache</value>
+        </interfaces>
+        <service-properties>
+            <entry key="alias" value="opennms.passiveRestConnectionHealthCheck"/>
+        </service-properties>
+    </service>
+
+</blueprint>
